Experimental Alert (#11)
This pull request adds a warning alert to the main application UI to
inform users that the demo is an experimental beta release and the code
is not audited. This helps set expectations and encourages users to
exercise caution.

UI/UX improvements:

* Added a warning `Alert` component to the top of the main page in
`App.tsx` to notify users about the experimental nature and unaudited
status of the demo.